requirements.txt
Flask
flask_classful
transformers
torch
langchain
langchain-community
langchain-huggingface
bitsandbytes
accelerate
langserve[all]
python-dotenv